[#436] Decoupled drive controls from rover node
