+++ robot/rospackages/src/mcu_control/CMakeLists.txt
@@ -19,6 +19,8 @@ find_package(catkin REQUIRED COMPONENTS
 
 ## Generate services in the 'srv' folder
 add_service_files(
+  DIRECTORY
+  srv
   FILES
   ## arm
   ArmRequest.srv
